requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
pycodestyle==2.8.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
tzlocal==2.1
urllib3==1.26.7
